Bring in enhancements from N9OZB's fork

+++ models/cost.cc
@@ -36,8 +36,8 @@ http://morse.colorado.edu/~tlen5510/text/classwebch3.html
 	int C = 3;		// 3dB for Urban
 	float lRxH = log10(11.75 * RxH);
 	float C_H = 3.2 * (lRxH * lRxH) - 4.97;	// Large city (conservative)
-	int c0 = 69.55;
-	int cf = 26.16;
+	float c0 = 69.55;
+	float cf = 26.16;
 	if (f > 1500) {
 		c0 = 46.3;
 		cf = 33.9;
